+++ graph/__init__.py
@@ -1,0 +1,6 @@
+"""Module for the graph used in pathfinding"""
+
+
+from .graph import KytosGraph
+
+__all__ = ('KytosGraph',)
