+++ graph/filters.py
@@ -1,0 +1,30 @@
+"""Filters for usage in pathfinding."""
+
+def lazy_filter(filter_type, filter_func):
+    """
+    Lazy typed filter on top of the built-in function.
+
+    It's meant to be used when the values to be filtered for
+    are only defined later on dynamically at runtime.
+    """
+
+    def filter_closure(value, items):
+        if not isinstance(value, filter_type):
+            raise TypeError(f"Expected type: {filter_type}")
+        return filter(filter_func(value), items)
+
+    return filter_closure
+
+def filter_le(metric):
+    """Lazy filter_le."""
+    return lambda x: (lambda nx_edge_tup: nx_edge_tup[2].get(metric, x) <= x)
+
+
+def filter_ge(metric):
+    """Lazy filter_ge."""
+    return lambda x: (lambda nx_edge_tup: nx_edge_tup[2].get(metric, x) >= x)
+
+
+def filter_in(metric):
+    """Lazy filter_in."""
+    return lambda x: (lambda nx_edge_tup: x in nx_edge_tup[2].get(metric, {x}))
