+++ tests/unit/graph/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for KytosGraph and helpers."""
